--- blocks/columns-new/columns-new.css
@@ -1,104 +1,0 @@
-.columns-new > div {
-    display: flex;
-    flex-direction: column;
-}
-
-.columns-new img {
-    width: 100%;
-}
-
-.columns-new h1 {
-    font-weight: 700;
-    line-height: 1.75rem;
-    font-size: var(--font-size-hl-1250);
-    color: var(--headline-color);
-    margin-bottom: var(--spacing-vertical-s);
-}
-
-.columns-new .copyright {
-    background-color: rgba(255, 255, 255, .6);
-    padding-top: var(--spacing-vertical-xxs);
-    padding-right: var(--spacing-horizontal-xs);
-    padding-bottom: 0.6rem;
-    padding-left: var(--spacing-horizontal-xs);
-    position: absolute;
-    right: 0;
-    bottom: 100%;
-}
-
-.columns-new .caption {
-    padding-top: var(--spacing-vertical-xs);
-}
-
-.columns-new .copyright-text::before {
-    content: "©";
-    padding-right: var(--spacing-horizontal-xxs)
-}
-
-.columns-new > div > div {
-    order: 1;
-}
-
-.columns-new > div > div:first-child {
-    font-size: var(--font-size-body-750);
-    margin-bottom: var(--spacing-vertical-m);
-    line-height: var(--line-height-caption-750);
-}
-
-.columns-new > div > div:first-child p {
-    position: relative;
-    display: flex;
-}
-
-.columns-new > div > div:last-child {
-    letter-spacing: 0.005em;
-    line-height: var(--line-height-body-1000);
-}
-
-.columns-new > div > div:last-child:not(:has(h1)) {
-    margin-top: var(--spacing-vertical-s);
-}
-
-.columns-new > div > div:last-child p {
-    padding-top: 1.5rem;
-}
-
-.columns-new > div > div:last-child p:first-of-type {
-    padding-top: unset;
-}
-
-.columns-new > div > .columns-img-col {
-    order: 0;
-}
-
-.columns-new > div > .columns-img-col img {
-    display: block;
-}
-
-.columns-new-container {
-    margin: calc(var(--spacing-vertical-l) + var(--spacing-vertical-xl)) auto;
-    padding-left: var(--section-margin);
-    padding-right: var(--section-margin);
-    max-width: 102rem;
-}
-
-@media (min-width: 48rem) {
-    .columns-new > div {
-        flex-direction: unset;
-        gap: 1.5rem;
-    }
-
-    .columns-new > div > div {
-        flex: 1;
-        order: unset;
-    }
-
-    .columns-new > div > div:first-child {
-        margin-bottom: unset;
-    }
-
-    .columns-new h1 {
-        line-height: 2.25rem;
-        font-size: var(--font-size-hl-2000);
-    }
-}
